requirements.txt
flask==3.0.0
requests==2.31.0
openai==1.3.0
anthropic==0.7.0
PyGithub==2.1.1
playwright==1.40.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
uuid7==0.1.0
pyyaml==6.0.1
jinja2==3.1.2
gunicorn==21.2.0
tabulate==0.9.0
